make modal close on changes saved
